+++ src/MnistReaderCore.h
@@ -35,7 +35,7 @@ class MnistReaderCore {
   //! Destructor
   virtual ~MnistReaderCore();
   //! MNISTデータのラベルデータ読み込み
-  std::vector<double> readMnistLabel(const std::string& filename, const bool& oneHotLabel);
+  std::vector<float> readMnistLabel(const std::string& filename, const bool& oneHotLabel);
   //! エンディアン変換
   int reverseInt (int data);
